requirements.txt
Django==5.1.3
djangorestframework==3.16.1
django-cors-headers==4.4.0
drf-spectacular==0.29.0
aiogram==3.12.0
pandas==2.2.3
openpyxl==3.1.5
python-docx==1.1.2
fpdf==1.7.2
pillow==11.0.0
requests==2.31.0
django-filter==24.2

# Production
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
python-dotenv==1.0.1
